Enhance graphics and refactor data structures
Updated `XcsbClient` initialization to create larger windows and added new drawing methods for enhanced graphical capabilities. Refactored `Arc` and `Point` structs to follow C# naming conventions and optimize property types. Implemented several methods in `XProto` to improve socket communication and memory management for graphical requests. Updated `using` directives to include response models for better handling of socket interactions.

+++ Src/Models/Point.cs
@@ -7,6 +7,6 @@ using System.Threading.Tasks;
 namespace Xcsb.Models;
 public struct Point
 {
-    public int X;
-    public int Y;
+    public ushort X;
+    public ushort Y;
 }
